// backend/src/common/dto/device-event.dto.ts
export type DeviceEventType =
  | 'OK'
  | 'WARNING'
  | 'ALARM'
  | 'DISCONNECTED'
  | 'RECONNECTED';

export interface DeviceEventDto {
  timestamp: string;
  machineType: string;
  instanceId: string;
  plantId: string;
  type: DeviceEventType;
  previousType?: DeviceEventType;
  message: string;
}